requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
pymysql==1.1.0
python-dotenv==1.0.0
Werkzeug==3.1.4
gunicorn==23.0.0
